pyproject.toml
[project]
name = "code_weaver"
version = "0.1.0"
description = "An AI Coding Agent built with LangGraph and LangChain"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain",
    "langgraph",
    "langchain-openai",
    "python-dotenv",
    "streamlit>=1.57.0",
    "pyyaml>=6.0.3",
    "pydantic>=2.13.4",
    "googlesearch-python>=1.3.0",
    "ruff>=0.15.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "tomli>=2.4.1",
]
